+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### v0.17 (unreleased)
 #### Changed
 - modified gaze calibration to save log, allow data streaming
+- Turandot transfers trial logs to PC along with main data file (if "Upload" option selected)
 #### Added 
 - manual audiometer
 #### Fixed